Refactor Win32 scanpath into platform abstraction layer
Move Windows-specific ctypes-based scanpath logic from app/utils/generic.py into app/utils/platform/windows.py to isolate OS-specific code.
